Move helper function get_experiment_dict() from apps/experiments/views.py to /apps/experiments/helper_functions.py
The views files should remain well structured and therefore only contain the routes.
